Updated UserRoles to use UserRole. Got rd of short and Role enums

+++ src/SFA.DAS.EmployerAccounts.ReadStore/SFA.DAS.EmployerAccounts.ReadStore/Queries/HasRoleQueryHandler.cs
@@ -23,7 +23,7 @@ namespace SFA.DAS.EmployerAccounts.ReadStore.Queries
                 .Where(r => r.UserRef == request.UserRef
                           && r.AccountId == request.EmployerAccountId).ToListAsync(cancellationToken);
 
-            var hasRole = userRoles.Any(r => r.Roles.Any(role => request.UserRoles.Any(requestRole => (short)requestRole == role)));
+            var hasRole = userRoles.Any(r => r.Roles.Any(role => request.UserRoles.Any(requestRole => requestRole == role)));
 
             return new HasRoleQueryResult{ HasRole = hasRole };
         }
